+++ src/main/java/com/komentum/seed/DevDataGenerator.java
@@ -8,8 +8,8 @@ import com.komentum.seed.seeder.ThemeBoardSeeder;
 import com.komentum.seed.seeder.ThemeComponentSeeder;
 import com.komentum.seed.seeder.UserSeeder;
 import com.komentum.theme.component.domain.DesignComponent;
-import com.komentum.theme.component.service.ColorStyleSeeder;
-import com.komentum.theme.component.service.ComponentTypeSeeder;
+import com.komentum.theme.component.service.seeder.ColorStyleSeeder;
+import com.komentum.theme.component.service.seeder.ComponentTypeSeeder;
 import com.komentum.theme.theme.domain.ThemeComponent;
 import com.komentum.user.domain.User;
 import jakarta.annotation.PostConstruct;
