+++ src/main/java/com/komentum/theme/component/controller/ColorStyleController.java
@@ -6,8 +6,8 @@ import com.komentum.theme.component.dto.ColorStyleResponse;
 import com.komentum.theme.component.dto.ColorStyleUpdateRequest;
 import com.komentum.theme.component.dto.SeedResult;
 import com.komentum.theme.component.mapper.ColorStyleMapper;
-import com.komentum.theme.component.service.ColorStyleSeeder;
 import com.komentum.theme.component.service.ColorStyleService;
+import com.komentum.theme.component.service.seeder.ColorStyleSeeder;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
 import java.util.List;
