+++ src/main/java/com/komentum/theme/component/controller/ComponentTypeController.java
@@ -6,8 +6,8 @@ import com.komentum.theme.component.dto.ComponentTypeDto;
 import com.komentum.theme.component.dto.ComponentTypeUpdateRequest;
 import com.komentum.theme.component.dto.SeedResult;
 import com.komentum.theme.component.mapper.ComponentTypeMapper;
-import com.komentum.theme.component.service.ComponentTypeSeeder;
 import com.komentum.theme.component.service.ComponentTypeService;
+import com.komentum.theme.component.service.seeder.ComponentTypeSeeder;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
 import java.util.List;
