+++ src/main/java/com/komentum/theme/component/service/seeder/AbstractMapJsonSeeder.java
@@ -1,4 +1,4 @@
-package com.komentum.theme.component.service;
+package com.komentum.theme.component.service.seeder;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.komentum.theme.utils.JsonUtils;
