+++ src/test/java/com/komentum/test/data/ThemeDataGenerator.java
@@ -7,8 +7,8 @@ import com.komentum.theme.component.domain.DesignComponent;
 import com.komentum.theme.component.repository.ColorStyleRepository;
 import com.komentum.theme.component.repository.ComponentTypeRepository;
 import com.komentum.theme.component.repository.DesignComponentRepository;
-import com.komentum.theme.component.service.ColorStyleSeeder;
-import com.komentum.theme.component.service.ComponentTypeSeeder;
+import com.komentum.theme.component.service.seeder.ColorStyleSeeder;
+import com.komentum.theme.component.service.seeder.ComponentTypeSeeder;
 import com.komentum.theme.theme.domain.ThemeComponent;
 import com.komentum.theme.theme.domain.ThemeImage;
 import com.komentum.theme.theme.domain.ThemeStyle;
